+++ src/main/java/net/brdle/collectorsreap/common/ForgeEvents.java
@@ -50,8 +50,11 @@ public class ForgeEvents {
 
 	@SubscribeEvent
 	public void onWanderingTrader(WandererTradesEvent e) {
-		if (CRConfig.verify(CRItems.LIME) && CRConfig.verify(CRItems.LIME_SEEDS)) {
+		if (CRConfig.verify(CRItems.LIME)) {
 			e.getGenericTrades().add((ent, r) -> new MerchantOffer(new ItemStack(Items.EMERALD, 1), Util.gs(CRItems.LIME_SEEDS), 5, 1, 1));
+		}
+		if (CRConfig.verify(CRItems.PINK_DRAGON_FRUIT)) {
+			e.getGenericTrades().add((ent, r) -> new MerchantOffer(new ItemStack(Items.EMERALD, 1), Util.gs(CRItems.DRAGON_FRUIT_SEEDS), 5, 1, 1));
 		}
 		if (CRConfig.verify(CRItems.PORTOBELLO)) {
 			e.getGenericTrades().add((ent, r) -> new MerchantOffer(new ItemStack(Items.BROWN_MUSHROOM, 4), Util.gs(CRItems.PORTOBELLO), 10, 1, 1));
@@ -59,7 +62,7 @@ public class ForgeEvents {
 	}
 
 	// Surge
-	@SubscribeEvent(priority = EventPriority.HIGHEST)
+	@SubscribeEvent(priority = EventPriority.HIGHEST) // Apply damage modifiers early
 	public void onSurgeDamage(LivingDamageEvent e) {
 		final MobEffect surge = CREffects.SURGE.get();
 		final DamageSource source = e.getSource();